+++ tests/Backend/Unit/Console/ImportGreenlight2Test.php
@@ -263,6 +263,11 @@ class ImportGreenlight2Test extends TestCase
         $this->assertEquals('012345', Room::find('hij-klm-xyz-123')->access_code);
         $this->assertNull(Room::find('hij-klm-xyz-234')->access_code);
 
+        // check expert mode is enabled for all rooms
+        foreach (Room::where('id', '!=', $existingRoom->id)->get() as $room) {
+            $this->assertTrue($room->expert_mode);
+        }
+
         // check room settings
         $this->assertFalse(Room::find('hij-klm-xyz-123')->mute_on_start);
         $this->assertFalse(Room::find('hij-klm-xyz-123')->everyone_can_start);